requirements.ultralight.txt
# Ultra-lightweight requirements for <300MB Docker image
# NO local ML models - API-based only
# NO PyTorch, NO sentence-transformers, NO heavy dependencies

# ==================== Core Framework ====================
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# ==================== Configuration ====================
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ==================== LlamaIndex Core (Minimal) ====================
llama-index-core==0.10.0

# ==================== LLM Providers (API-based only) ====================
llama-index-llms-groq==0.1.3
# Uncomment if using:
# llama-index-llms-openai==0.1.5

# ==================== Embeddings (API-based ONLY - no local models!) ====================
# Use OpenAI or Cohere API for embeddings (no local models)
llama-index-embeddings-openai==0.1.5
# OR use Cohere:
# llama-index-embeddings-cohere==0.1.5

# ==================== Document Processing (Minimal) ====================
pypdf==3.17.1
python-docx==1.1.0
beautifulsoup4==4.12.2
lxml==4.9.3

# ==================== Web & HTTP ====================
requests==2.31.0

# ==================== Token Management ====================
tiktoken==0.5.2

# ==================== AWS ====================
boto3==1.34.0

# ==================== Authentication ====================
PyJWT==2.8.0
cryptography==41.0.7

# ==================== Utilities ====================
python-multipart==0.0.6

# ==================== Performance ====================
orjson==3.9.10

# ==================== Vector Stores (Optional - lightweight only) ====================
# For Qdrant (lightweight client):
# qdrant-client==1.7.0  # ~10MB
# llama-index-vector-stores-qdrant==0.2.0  # ~5MB

# ==================== EXCLUDED (Too Heavy) ====================
# ❌ sentence-transformers (800MB+ with PyTorch)
# ❌ llama-index-embeddings-huggingface (depends on transformers)
# ❌ torch/pytorch (800MB+)
# ❌ transformers (200MB+)
# ❌ aiohttp (not needed for basic API calls)
# ❌ trafilatura (web scraping - use requests instead)

# Total size: ~150-200MB (base alpine ~50MB + packages ~100-150MB)
# Final image: <300MB ✅
